WIP: Giving up on bundling
Every approach tried using Stackblitz's bunlder, or using Vercel NCC, or loading packages from the example server, failed with webcontainers.io.
The only way we can get this working at present is with `npm install @ably/cli` as part of the container bootstrap, which means we have to wait aorund 30s for the container to become available, so not workable.